+++ server/src/backend/ReferenceProvider.ts
@@ -1,52 +1,80 @@
+import * as fs from "fs";
 import { Position, TextDocument } from "vscode-languageserver-textdocument";
-import { LpcFacade } from "./facade";
-import { Location } from "vscode-languageserver";
+import { LpcFacade, getProjectFiles } from "./facade";
+import { CancellationToken, Location } from "vscode-languageserver";
 import {
-    firstEntry,
     getFilenameForSymbol,
     lexRangeFromContext,
     lexRangeToLspRange,
-    rangeFromTokens,
 } from "../utils";
 import {
-    IReferenceableSymbol,
-    getSymbolsOfTypeSync,
     isInstanceOfIReferenceSymbol,
     isInstanceOfIReferenceableSymbol,
 } from "../symbols/base";
-import { ContextSymbolTable } from "./ContextSymbolTable";
+
 import { ParserRuleContext } from "antlr4ng";
 import { LPCToken } from "../parser3/LPCToken";
-import { BaseSymbol } from "antlr4-c3";
+import { BaseSymbol, ScopedSymbol } from "antlr4-c3";
 import { URI } from "vscode-uri";
-import {
-    FunctionHeaderContext,
-    FunctionHeaderDeclarationContext,
-} from "../parser3/LPCParser";
+import { FunctionHeaderContext } from "../parser3/LPCParser";
 import { LpcBaseMethodSymbol } from "../symbols/methodSymbol";
+import { isIdentifierPart } from "./scanner-utils";
 
+type CandidatePosition = {
+    filename: string;
+    position: number;
+};
 export class ReferenceProvider {
     constructor(private backend: LpcFacade) {}
 
     public async handleReferenceRequest(
         doc: TextDocument,
-        position: Position
+        position: Position,
+        token: CancellationToken = CancellationToken.None
     ): Promise<Location[]> {
         const docFilename = URI.parse(doc.uri).fsPath;
-
-        // NTBLA: decide if we should automatically parse all or not
-        //await this.backend.parseAllIfNeeded();
-
         const results: Location[] = [];
         const seen: Set<BaseSymbol> = new Set();
 
         const sym = this.backend.symbolContainingPosition(doc.uri, position);
-
         if (!sym) {
             return [];
         }
 
-        // collect possible references - they will be scanned later to confirm
+        const referenceName = sym.name;
+
+        // scan all files for candidates
+        // TODO: implemenet scope check so that we don't do a full scan for things like block-scoped variables
+        const projectFiles = await getProjectFiles(
+            this.backend.workspaceDir,
+            token
+        );
+        const candidatePositions: CandidatePosition[] = [];
+        for (const file of projectFiles) {
+            if (token.isCancellationRequested) break;
+
+            const readPromise = new Promise<void>((resolve, reject) => {
+                fs.readFile(file, "utf-8", (err, text) => {
+                    const positions = searchForName(text, referenceName);
+                    candidatePositions.push(
+                        ...positions.map((p) => ({
+                            filename: file,
+                            position: p,
+                        }))
+                    );
+                    resolve();
+                });
+            });
+
+            await readPromise;
+        }
+
+        // TODO: cache this
+        const candidateFiles = new Set(
+            candidatePositions.map((cp) => cp.filename)
+        );
+
+        // add the symbol and its reference to the list
         const refsToScan: BaseSymbol[] = [sym];
         if (isInstanceOfIReferenceSymbol(sym))
             refsToScan.push(sym.getReference());
@@ -58,16 +86,19 @@ export class ReferenceProvider {
             symFile, // the current file
             ...(this.backend.includeRefs.get(symFile) ?? []), // any included files
             ...(this.backend.fileRefs.get(symFile) ?? []), // any files that reference this one
+            ...(candidateFiles ?? []),
         ];
 
         // now scan the files
         const seenFiles = new Set<string>();
         for (const file of files) {
+            if (token.isCancellationRequested) break;
             if (seenFiles.has(file)) continue;
             seenFiles.add(file);
 
-            // TODO: this might leave files in memory that aren't needed
+            // load files and evaluate them so that any arrow symbols will be resolved
             const refCtx = this.backend.loadLpc(file);
+            refCtx.evaluateProgram();
 
             // look for name matches and make sure their references are filled in.
             const { symbolTable } = refCtx;
@@ -79,7 +110,7 @@ export class ReferenceProvider {
                 }
             });
 
-            // now find the "definition" symbol and add any its file
+            // find the "definition" symbol and add any its file
             const refSym = await refCtx.symbolTable.resolve(sym.name, true);
             if (
                 isInstanceOfIReferenceableSymbol(refSym) &&
@@ -94,13 +125,28 @@ export class ReferenceProvider {
             files.push(...(this.backend.fileRefs.get(file) ?? []));
         }
 
-        while (refsToScan.length > 0) {
-            const ref = refsToScan.pop();
+        // now, the original symbol is either a reference or referencable.
+        let definitionSymbol: BaseSymbol;
+        if (isInstanceOfIReferenceSymbol(sym)) {
+            definitionSymbol = sym.getReference();
+        } else if (isInstanceOfIReferenceableSymbol(sym)) {
+            definitionSymbol = sym;
+        } else {
+            // this should never happen
+            console.log("Unexpected symbol type in reference provider");
+        }
+
+        const finalList = [definitionSymbol];
+
+        while (finalList.length > 0) {
+            if (token.isCancellationRequested) break;
+
+            const ref = finalList.pop();
             if (!ref || seen.has(ref)) continue;
             seen.add(ref);
 
             if (isInstanceOfIReferenceableSymbol(ref)) {
-                refsToScan.push(...ref.references);
+                finalList.push(...ref.references);
             }
 
             let parseInfo = ref.context as ParserRuleContext;
@@ -113,8 +159,8 @@ export class ReferenceProvider {
                     parseInfo = header._functionName;
                 }
             }
-            const token = parseInfo.start as LPCToken;
-            const filename = token.filename;
+            const t = parseInfo.start as LPCToken;
+            const filename = t.filename;
             const range = lexRangeToLspRange(lexRangeFromContext(parseInfo));
 
             results.push({
@@ -123,6 +169,68 @@ export class ReferenceProvider {
             });
         }
 
+        // release candidate files
+        for (const candidateFile of candidateFiles) {
+            this.backend.releaseLpc(candidateFile);
+        }
+
+        if (token.isCancellationRequested) {
+            console.log("Reference provider cancelled");
+        }
+
         return results;
     }
+}
+
+/**
+ * Determine the smallest scope in which a symbol may have references.
+ * @param symbol
+ * @returns undefeind if the scope cannot be determined,
+ * which means a reference can be anywhere
+ */
+function getSymbolScope(symbol: BaseSymbol): ScopedSymbol | undefined {
+    // TODO: implement this
+    return undefined;
+}
+
+/**
+ * Searches a file for instances of a symbol name
+ * @param sourceText the source code of the file to scan
+ * @param search
+ * @returns an array of starting positions of the symbol name
+ */
+function searchForName(
+    sourceText: string,
+    search: string,
+    containerStart: number = 0,
+    containerEnd: number = sourceText.length
+): readonly number[] {
+    const positions: number[] = [];
+    if (!search || !search.length) positions;
+
+    const sourceLength = sourceText.length;
+    const symbolNameLength = search.length;
+
+    let position = sourceText.indexOf(search, containerStart);
+    while (position >= 0) {
+        // If we are past the end, stop looking
+        if (position > containerEnd) break;
+
+        // We found a match.  Make sure it's not part of a larger word (i.e. the char
+        // before and after it have to be a non-identifier char).
+        const endPosition = position + symbolNameLength;
+
+        if (
+            (position === 0 ||
+                !isIdentifierPart(sourceText.charCodeAt(position - 1))) &&
+            (endPosition === sourceLength ||
+                !isIdentifierPart(sourceText.charCodeAt(endPosition)))
+        ) {
+            // Found a real match.  Keep searching.
+            positions.push(position);
+        }
+        position = sourceText.indexOf(search, position + symbolNameLength + 1);
+    }
+
+    return positions;
 }